+++ Alicargo.Core/Helpers/CalculationDataHelper.cs
@@ -4,11 +4,11 @@ namespace Alicargo.Core.Helpers
 {
 	public static class CalculationDataHelper
 	{
-		public static decimal GetMoney(CalculationData calculation, decimal insuranceRate)
+		public static decimal GetMoney(CalculationData calculation, float insuranceRate)
 		{
 			return (decimal)calculation.Weight * calculation.TariffPerKg
 				   + calculation.ScotchCost
-				   + calculation.Value * insuranceRate
+				   + calculation.Value * (decimal)insuranceRate
 				   + calculation.FactureCost
 				   + calculation.FactureCostEx
 				   + calculation.TransitCost
