+++ Alicargo.DataAccess.Contracts/Contracts/Application/ApplicationExtendedData.cs
@@ -128,9 +128,9 @@ namespace Alicargo.DataAccess.Contracts.Contracts.Application
 
 		public decimal? PickupCost { get; set; }
 
-		public decimal InsuranceRate { get; set; }
+		public float  InsuranceRate { get; set; }
 
-		public decimal InsuranceRateForClient { get; set; }
+		public float InsuranceRateForClient { get; set; }
 
 
 		public long SenderId { get; set; }
