+++ Alicargo.DataAccess.Contracts/Repositories/Application/IApplicationRepository.cs
@@ -18,7 +18,7 @@ namespace Alicargo.DataAccess.Contracts.Repositories.Application
 
 		long GetClientId(long id);
 
-		decimal GetDefaultInsuranceRate();
+		float GetDefaultInsuranceRate();
 
 		ApplicationExtendedData GetExtendedData(long id);
